+++ MediViewer/src/main/java/com/tjoeun/mediviewer/domain/WorkList.java
@@ -1,0 +1,14 @@
+package com.tjoeun.mediviewer.domain;
+
+public interface WorkList {
+	String getPid();
+	String getPname();
+	Integer getStudyKey();
+	String getModality();
+	String getStudyDesc();
+	String getStudyDate();
+	String getReportStatus();
+	Integer getSeriesCnt();
+	Integer getImageCnt();
+}
+
