+++ MediViewer/src/main/java/com/tjoeun/mediviewer/service/SeriesService.java
@@ -1,0 +1,8 @@
+package com.tjoeun.mediviewer.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SeriesService {
+
+}
