+++ MediViewer/src/main/resources/static/js/commonAxios.js
@@ -1,0 +1,9 @@
+const getDefList = () => {
+	axios.get("/api/lists")
+		.then(response => {
+			const rData = response.data;
+			setCount(count);
+			setListItems(rData.items);
+		})
+		.catch(error => console.error(error));
+}
