--- demo/src/main/java/com/study/medi/domain/Comment.java
@@ -1,41 +1,0 @@
-package com.study.medi.domain;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import com.study.medi.domain.ImageTab.ImageId;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Table(name="COMMENT")
-public class Comment {
-	
-    @Id
-    @Column(name = "STUDYKEY")
-    private int studyKey;
-    
-    @Column(name = "COMMENT")
-    private String comment;
-    
-	@LastModifiedDate
-	@Column(name="MODIFIED_DATE", columnDefinition="DATE DEFAULT SYSDATE")
-	private LocalDateTime modifiedDate;
-	
-	@CreatedDate
-	@Column(name="CREATED_DATE")
-	private LocalDateTime createdDate;
-	
-}
