--- demo/src/main/java/com/study/medi/domain/WorkList.java
@@ -1,14 +1,0 @@
-package com.study.medi.domain;
-
-public interface WorkList {
-	String getPid();
-	String getPName();
-	String getModality();
-	String getStudyDesc();
-	String getStudyDate();
-	String getReportStatus();
-	Integer getSeriesCnt();
-	Integer getImageCnt();
-	
-}
-
