--- demo/src/main/java/com/study/medi/service/CommentService.java
@@ -1,8 +1,0 @@
-package com.study.medi.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CommentService {
-
-}
