--- demo/src/main/java/com/study/medi/service/SeriesService.java
@@ -1,8 +1,0 @@
-package com.study.medi.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class SeriesService {
-
-}
