--- demo/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-spring.application.name=demo
-
-# log4jdbc
-spring.datasource.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-spring.datasource.url=jdbc:log4jdbc:oracle:thin:@localhost:1521/xe
-spring.datasource.username=dicom
-spring.datasource.password=1234
-spring.datasource.hikari.maximum-pool-size=5
-
-# JPA hivernatei
-spring.jpa.database-platform=org.hibernate.dialect.OracleDialect
-spring.jpa.open-in-view=false
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.hbm2ddl.auto=update
