--- demo/src/main/resources/properties/env.properties
@@ -1,9 +1,0 @@
-# Java \uD30C\uC77C\uC5D0\uC11C \uD14C\uC2A4\uD2B8
-java.file.test=Hello, World!! 
-
-# properties \uD30C\uC77C\uC5D0\uC11C \uD14C\uC2A4\uD2B8
-properties.file.port=8080
-
-# img file Path
-#dicome.img.path ="C:\\Users\\TJ\\Desktop\\gjPJ\\common"
-dicome.img.path =${HOME}\Desktop\img
