--- demo/src/main/resources/templates/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-<script src="https://unpkg.com/cornerstone-core"></script>
-<script src="https://unpkg.com/cornerstone-tools"></script>
-<script src="https://unpkg.com/dicom-parser"></script>
-<script src="https://unpkg.com/cornerstone-wado-image-loader"></script>
-</head>
-<body>
-	<h2>DICOM IMAGE VIEWER</h2>
-	<body>
-    <h1 th:text="${envImgPath}">Default Title</h1>
-	<div id="dicomImage"></div>
-	
-  	<script type="text/javascript">
-        // JavaScript에서 Thymeleaf 변수를 사용
-        
-        var envImgPath = '[[${envImgPath}]]';
-        console.log("envImgPath : ", envImgPath) ;
-    </script>
-	
-	<script src="app.js"></script>
-
-</body>
-</html>
